rename eventQueue to CustomRoleQueue
It's a much more accurate name for what the queue is doing.
